refactor: split managers & fields from models in `users` app
