+++ src/users/managers.py
@@ -1,0 +1,33 @@
+from django.contrib.auth.models import BaseUserManager
+
+
+
+class UserManager(BaseUserManager):
+
+    def create_user(self, phone, password, **other_fields):
+        if phone is None:
+            raise ValueError("Phone not given")
+
+        user = self.model(phone=phone, **other_fields)
+        user.set_password(password)
+        user.save(using=self._db)
+        return user
+
+    def create_superuser(self, phone, password=None, **other_fields):
+        other_fields.setdefault('is_staff', True)
+        other_fields.setdefault('is_superuser', True)
+
+        if other_fields.get('is_staff') is not True:
+            raise ValueError('Superuser must have is_staff=True.')
+        if other_fields.get('is_superuser') is not True:
+            raise ValueError('Superuser must have is_superuser=True.')
+
+        return self.create_user(phone, password, **other_fields)
+
+
+class ClientManager(BaseUserManager):
+    def create(self, phone, **other_fields):
+        user = self.model(phone=phone, **other_fields)
+        user.set_unusable_password()  # XXX: this can also be a random string password
+        user.save(using=self._db)
+        return user
