+++ src/users/models.py
@@ -1,53 +1,13 @@
 from django.db import models
 from django.contrib.auth.models import (
     AbstractBaseUser,
-    BaseUserManager,
     PermissionsMixin
 )
-from django.core.exceptions import ValidationError
 
 from core.models import BaseModel
 from core.validators import phone_validator
-
-
-
-class UserManager(BaseUserManager):
-
-    def create_user(self, phone, password, **other_fields):
-        if phone is None:
-            raise ValueError("Phone not given")
-
-        user = self.model(phone=phone, **other_fields)
-        user.set_password(password)
-        user.save(using=self._db)
-        return user
-
-    def create_superuser(self, phone, password=None, **other_fields):
-        other_fields.setdefault('is_staff', True)
-        other_fields.setdefault('is_superuser', True)
-
-        if other_fields.get('is_staff') is not True:
-            raise ValueError('Superuser must have is_staff=True.')
-        if other_fields.get('is_superuser') is not True:
-            raise ValueError('Superuser must have is_superuser=True.')
-
-        return self.create_user(phone, password, **other_fields)
-
-
-class PhoneNumberField(models.CharField):
-    def get_prep_value(self, value):
-        if value is None:
-            return value
-
-        try:
-            regex = phone_validator(value)
-        except ValidationError:
-            raise
-
-        phone_parts = regex.groupdict()
-        phone = phone_parts["operator"]+phone_parts["middle3"]+phone_parts["last4"]
-        return phone
-
+from .managers import UserManager, ClientManager
+from .fields import PhoneNumberField
 
 
 class User(AbstractBaseUser, PermissionsMixin, BaseModel):
@@ -72,14 +32,6 @@ class Staff(User):
     class Meta:
         proxy = True
 
-
-
-class ClientManager(BaseUserManager):
-    def create(self, phone, **other_fields):
-        user = self.model(phone=phone, **other_fields)
-        user.set_unusable_password()  # XXX: this can also be a random string password
-        user.save(using=self._db)
-        return user
 
 class Client(User):
     objects = ClientManager()
